requirements.txt
msgpack==1.0.0
Pillow==7.2.0
qrcode==6.1
six==1.15.0
